Replace 4x4 legs, add tool port
Replace the 4x4 legs with crossbraced 2x4s that will hopefully be
sturdier. Add a "port" in the center of one of the benches to allow for
swappable tools (mitre saw, bandsaw, etc). Add cross-support beams from
one table piece to the other to avoid having a table leg right where
I'll be sitting.